+++ tests/models/test_filter.py
@@ -1,6 +1,5 @@
 from mex.common.models import (
     EXTRACTED_MODEL_CLASSES,
-    EXTRACTED_MODEL_CLASSES_BY_NAME,
     FILTER_MODEL_CLASSES,
     FilterField,
     PersonFilter,
@@ -14,19 +13,12 @@ def test_all_filter_classes_are_defined() -> None:
 
 def test_all_filter_fields_are_defined() -> None:
     for filter_cls in FILTER_MODEL_CLASSES:
-        extracted_cls = EXTRACTED_MODEL_CLASSES_BY_NAME[
-            f"Extracted{filter_cls.stemType}"
-        ]
-        assert set(filter_cls.model_fields) == set(extracted_cls.model_fields)
         field_defs = {
             field_name: (field_info.annotation, field_info.default)
             for field_name, field_info in filter_cls.model_fields.items()
             if field_name != "entityType"
         }
-        assert all(
-            (annotation, default) == (list[FilterField], [])
-            for annotation, default in field_defs.values()
-        )
+        assert field_defs == {"fields": (list[FilterField], [])}
 
 
 def test_filter_model_schema() -> None:
@@ -34,7 +26,7 @@ def test_filter_model_schema() -> None:
         "$defs": {
             "FilterField": {
                 "additionalProperties": False,
-                "description": "Entity filter field model.",
+                "description": "Filter definition for one field in the primary source.",
                 "properties": {
                     "fieldInPrimarySource": {
                         "anyOf": [{"type": "string"}, {"type": "null"}],
@@ -54,10 +46,10 @@ def test_filter_model_schema() -> None:
                         "default": None,
                         "title": "examplesInPrimarySource",
                     },
-                    "mappingRules": {
+                    "filterRules": {
                         "items": {"$ref": "#/$defs/FilterRule"},
                         "minItems": 1,
-                        "title": "mappingRules",
+                        "title": "filterRules",
                         "type": "array",
                     },
                     "comment": {
@@ -66,13 +58,13 @@ def test_filter_model_schema() -> None:
                         "title": "comment",
                     },
                 },
-                "required": ["mappingRules"],
+                "required": ["filterRules"],
                 "title": "FilterField",
                 "type": "object",
             },
             "FilterRule": {
                 "additionalProperties": False,
-                "description": "Entity filter rule model.",
+                "description": "A single filter rule to apply.",
                 "properties": {
                     "forValues": {
                         "anyOf": [
@@ -102,64 +94,10 @@ def test_filter_model_schema() -> None:
                 "title": "$Type",
                 "type": "string",
             },
-            "hadPrimarySource": {
+            "fields": {
                 "default": [],
                 "items": {"$ref": "#/$defs/FilterField"},
-                "title": "Hadprimarysource",
-                "type": "array",
-            },
-            "identifierInPrimarySource": {
-                "default": [],
-                "items": {"$ref": "#/$defs/FilterField"},
-                "title": "Identifierinprimarysource",
-                "type": "array",
-            },
-            "affiliation": {
-                "default": [],
-                "items": {"$ref": "#/$defs/FilterField"},
-                "title": "Affiliation",
-                "type": "array",
-            },
-            "email": {
-                "default": [],
-                "items": {"$ref": "#/$defs/FilterField"},
-                "title": "Email",
-                "type": "array",
-            },
-            "familyName": {
-                "default": [],
-                "items": {"$ref": "#/$defs/FilterField"},
-                "title": "Familyname",
-                "type": "array",
-            },
-            "fullName": {
-                "default": [],
-                "items": {"$ref": "#/$defs/FilterField"},
-                "title": "Fullname",
-                "type": "array",
-            },
-            "givenName": {
-                "default": [],
-                "items": {"$ref": "#/$defs/FilterField"},
-                "title": "Givenname",
-                "type": "array",
-            },
-            "isniId": {
-                "default": [],
-                "items": {"$ref": "#/$defs/FilterField"},
-                "title": "Isniid",
-                "type": "array",
-            },
-            "memberOf": {
-                "default": [],
-                "items": {"$ref": "#/$defs/FilterField"},
-                "title": "Memberof",
-                "type": "array",
-            },
-            "orcidId": {
-                "default": [],
-                "items": {"$ref": "#/$defs/FilterField"},
-                "title": "Orcidid",
+                "title": "fields",
                 "type": "array",
             },
         },